requirements.txt
streamlit
fpdf
Pillow